common: rename lib dir to src
