+++ packages/common/src/schema.ts
