+++ packages/common/src/util/inspect.ts
